+++ app/src/main/java/com/github/dfa/diaspora_android/ui/theme/ThemedCheckBoxPreference.java
@@ -41,7 +41,7 @@ public class ThemedCheckBoxPreference extends CheckBoxPreference implements Them
 
     @Override
     public void setColors() {
-        CheckBox checkBox = (CheckBox) rootLayout.findViewById(android.R.id.checkbox);
+        CheckBox checkBox = rootLayout.findViewById(android.R.id.checkbox);
         ThemeHelper.getInstance(AppSettings.get());
         ThemeHelper.updateCheckBoxColor(checkBox);
     }