+++ app/src/main/java/com/github/dfa/diaspora_android/util/ActivityUtils.java
@@ -21,7 +21,7 @@ import java.util.Locale;
 
 @SuppressWarnings({"WeakerAccess", "unused", "SameParameterValue"})
 public class ActivityUtils extends net.gsantner.opoc.util.ActivityUtils {
-    protected ActivityUtils(Activity activity) {
+    public ActivityUtils(Activity activity) {
         super(activity);
     }
 
@@ -77,10 +77,10 @@ public class ActivityUtils extends net.gsantner.opoc.util.ActivityUtils {
 
     /**
      * This method creates file sharing uri by using FileProvider
+     *
      * @return
      */
-    public static Uri getFileSharingUri(Context context,File file) {
-
-        return FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID,file);
+    public static Uri getFileSharingUri(Context context, File file) {
+        return FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID, file);
     }
 }