+++ app/src/main/java/net/gsantner/opoc/preference/nonsupport/LanguagePreference.java
@@ -1,14 +1,13 @@
-/*
- * ------------------------------------------------------------------------------
- * Gregor Santner <gsantner.net> wrote this. You can do whatever you want
- * with it. If we meet some day, and you think it is worth it, you can buy me a
- * coke in return. Provided as is without any kind of warranty. Do not blame or
- * sue me if something goes wrong. No attribution required.    - Gregor Santner
+/*#######################################################
  *
- * License: Creative Commons Zero (CC0 1.0)
- *  http://creativecommons.org/publicdomain/zero/1.0/
- * ----------------------------------------------------------------------------
- */
+ *   Maintained by Gregor Santner, 2017-
+ *   https://gsantner.net/
+ *
+ *   License: Apache 2.0
+ *  https://github.com/gsantner/opoc/#licensing
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+#########################################################*/
 
 /*
  * Requires:
@@ -49,8 +48,8 @@ package net.gsantner.opoc.preference.nonsupport;
 import android.annotation.TargetApi;
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.Nullable;
 import android.preference.ListPreference;
+import android.support.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 
@@ -114,15 +113,13 @@ public class LanguagePreference extends ListPreference {
         // Fetch readable details
         ContextUtils contextUtils = new ContextUtils(context);
         List<String> languages = new ArrayList<>();
-        Object bcof = contextUtils.getBuildConfigValue("APPLICATION_LANGUAGES");
+        Object bcof = contextUtils.getBuildConfigValue("DETECTED_ANDROID_LOCALES");
         if (bcof instanceof String[]) {
             for (String langId : (String[]) bcof) {
                 Locale locale = contextUtils.getLocaleByAndroidCode(langId);
                 languages.add(summarizeLocale(locale, langId) + ";" + langId);
             }
         }
-        String aa = contextUtils.str("build_conf__detected_languages");
-        String a = aa;
 
         // Sort languages naturally
         Collections.sort(languages);